requirements.txt
gym
numpy
pytorch
tensorboard
torchvision
opencv-python